requirements.txt
tinydb==4.5.0
feedparser==6.0.8
click==8.0.1
pyyaml==5.4.1
python-telegram-bot